+++ R/annotate_coess.R
@@ -5,6 +5,8 @@
 #' 
 #' @param input_ess data frame, A data frame output from `coessential_map()`, Default: NULL
 #' @param input_inflec data frame, A data frame output from get_inflection_points(), Default: NULL
+#' @param top_n vector, The number of top/bottom candidates to select (eg. if 10 there will be top top 
+#' co-essential and 10 anti-essential genes), Default: NULL
 #' 
 #' @return The input data frame containing co-essential correlation coefficients will be annotated with an additional column 
 #' `Candidate_gene` to indicate whether the gene is considered to a possible co-essential gene.
@@ -37,7 +39,7 @@
 #' @importFrom dplyr mutate filter pull rename arrange case_when
 #' @importFrom tibble tibble
 
-annotate_coess <- function(input_ess = NULL, input_inflec = NULL) {
+annotate_coess <- function(input_ess = NULL, input_inflec = NULL, top_n = NULL) {
   # Checkpoint
   if (is.null(input_ess)) {
     stop("No coessential dataframe found!")
@@ -49,13 +51,42 @@ annotate_coess <- function(input_ess = NULL, input_inflec = NULL) {
     stop("Input is not a dataframe, please check the input")
   }
   
-  res <- input_ess %>%
-    dplyr::arrange(-.data$Rank) %>%
-    dplyr::mutate(Candidate_gene = dplyr::case_when((.data$Padj_BH <
-                                                       0.05) & (.data$Rank >= input_inflec$Inflection_point_pos_byRank) ~
-                                                      TRUE, (.data$Padj_BH < 0.05) & (.data$Rank <=
-                                                                                        input_inflec$Inflection_point_neg_byRank) ~
-                                                      TRUE, TRUE ~ FALSE))
+  All_res <- NULL
+  for(g in seq_len(length(unique(input_ess$GeneNameID_A)))){
+    # g <- 1
+    gene <- unique(input_ess$GeneNameID_A)[g]
+    
+    select_input_ess <- input_ess %>%
+      dplyr::filter(.data$GeneNameID_A %in% gene)
+    
+    select_input_inflec <- input_inflec %>%
+      dplyr::filter(.data$GeneNameID_A %in% gene)
+    
+    res <- select_input_ess %>%
+      dplyr::arrange(-.data$Rank) %>%
+      dplyr::mutate(Candidate_gene = dplyr::case_when(
+        (.data$Padj_BH < 0.05) & (.data$Rank >= select_input_inflec$Inflection_point_pos_byRank) ~ TRUE,
+        (.data$Padj_BH < 0.05) & (.data$Rank <= select_input_inflec$Inflection_point_neg_byRank) ~ TRUE,
+        TRUE ~ FALSE
+      ))
+    
+    All_res <- dplyr::bind_rows(All_res, res)
+  }
   
-  return(res)
+  candidate <- All_res %>% dplyr::filter(.data$Candidate_gene)
+  top <- top_n
+  middle <- 18333 - (top_n *2)
+  if(length(unique(input_ess$GeneNameID_A)) == nrow(candidate)){
+    if (is.null(top_n)) {
+      stop("No candidates... Please add a number for top candidates")
+    } else {
+      message("No candidates... Manually selecting top ", top_n, " candidates.")
+      
+    }
+    All_res <- All_res %>% mutate(
+      Candidate_gene = rep(c(rep(TRUE, top), rep(FALSE, middle), rep(TRUE, top)), length(unique(input_ess$GeneNameID_A)))
+    )
+  }
+  
+  return(All_res)
 }